create ui-form for server_main and test it
