--- mainapp/server_mainwindow.py
@@ -1,129 +1,0 @@
-"""
-File contents structure of main window for server
-graphical presence. Build by PyQtDesigner and then
-edited by author. This work was complited during
-educated DB_and_PyQt course lesson 4 by GeekBrains,
-Moscow, November 2022.
-Maksim_Sapunov, Jenny6199@yandex.ru
-"""
-# from PyQt5 import QtWebKitWidgets
-from PyQt5 import QtCore, QtWidgets
-from PyQt5.QtWidgets import QApplication
-import sys
-
-
-class Ui_ServerMainWindow(QtWidgets.QWidget):
-    def setupUi(self, ServerMainWindow):
-        ServerMainWindow.setObjectName("ServerMainWindow")
-        ServerMainWindow.resize(635, 592)
-        self.centralwidget = QtWidgets.QWidget(ServerMainWindow)
-        self.centralwidget.setObjectName("centralwidget")
-        self.verticalWidget = QtWidgets.QWidget(self.centralwidget)
-        self.verticalWidget.setGeometry(QtCore.QRect(10, 70, 611, 271))
-        self.verticalWidget.setObjectName("verticalWidget")
-        self.verticalLayout = QtWidgets.QVBoxLayout(self.verticalWidget)
-        self.verticalLayout.setContentsMargins(0, 0, 0, 0)
-        self.verticalLayout.setObjectName("verticalLayout")
-        self.label = QtWidgets.QLabel(self.verticalWidget)
-        self.label.setObjectName("label")
-        self.verticalLayout.addWidget(self.label)
-        self.tableView = QtWidgets.QTableView(self.verticalWidget)
-        self.tableView.setEnabled(True)
-        self.tableView.setObjectName("tableView")
-        self.verticalLayout.addWidget(self.tableView)
-        self.horizontalLayoutWidget = QtWidgets.QWidget(self.centralwidget)
-        self.horizontalLayoutWidget.setGeometry(QtCore.QRect(10, 370, 301, 171))
-        self.horizontalLayoutWidget.setObjectName("horizontalLayoutWidget")
-        self.horizontalLayout = QtWidgets.QHBoxLayout(self.horizontalLayoutWidget)
-        self.horizontalLayout.setContentsMargins(0, 0, 0, 0)
-        self.horizontalLayout.setObjectName("horizontalLayout")
-        self.graphicsView = QtWidgets.QGraphicsView(self.horizontalLayoutWidget)
-        self.graphicsView.setObjectName("graphicsView")
-        self.horizontalLayout.addWidget(self.graphicsView)
-        self.horizontalLayoutWidget_2 = QtWidgets.QWidget(self.centralwidget)
-        self.horizontalLayoutWidget_2.setGeometry(QtCore.QRect(320, 370, 301, 171))
-        self.horizontalLayoutWidget_2.setObjectName("horizontalLayoutWidget_2")
-        self.horizontalLayout_2 = QtWidgets.QHBoxLayout(self.horizontalLayoutWidget_2)
-        self.horizontalLayout_2.setContentsMargins(0, 0, 0, 0)
-        self.horizontalLayout_2.setObjectName("horizontalLayout_2")
-        self.webView = QtWebKitWidgets.QWebView(self.horizontalLayoutWidget_2)
-        self.webView.setUrl(QtCore.QUrl("about:blank"))
-        self.webView.setObjectName("webView")
-        self.horizontalLayout_2.addWidget(self.webView)
-        self.line = QtWidgets.QFrame(self.centralwidget)
-        self.line.setGeometry(QtCore.QRect(-10, 50, 631, 16))
-        self.line.setFrameShape(QtWidgets.QFrame.HLine)
-        self.line.setFrameShadow(QtWidgets.QFrame.Sunken)
-        self.line.setObjectName("line")
-        self.line_2 = QtWidgets.QFrame(self.centralwidget)
-        self.line_2.setGeometry(QtCore.QRect(10, 340, 611, 16))
-        self.line_2.setFrameShape(QtWidgets.QFrame.HLine)
-        self.line_2.setFrameShadow(QtWidgets.QFrame.Sunken)
-        self.line_2.setObjectName("line_2")
-        self.label_2 = QtWidgets.QLabel(self.centralwidget)
-        self.label_2.setGeometry(QtCore.QRect(10, 350, 67, 17))
-        self.label_2.setObjectName("label_2")
-        self.label_3 = QtWidgets.QLabel(self.centralwidget)
-        self.label_3.setGeometry(QtCore.QRect(320, 350, 71, 20))
-        self.label_3.setObjectName("label_3")
-        self.horizontalLayoutWidget_3 = QtWidgets.QWidget(self.centralwidget)
-        self.horizontalLayoutWidget_3.setGeometry(QtCore.QRect(10, 10, 131, 41))
-        self.horizontalLayoutWidget_3.setObjectName("horizontalLayoutWidget_3")
-        self.horizontalLayout_3 = QtWidgets.QHBoxLayout(self.horizontalLayoutWidget_3)
-        self.horizontalLayout_3.setContentsMargins(0, 0, 0, 0)
-        self.horizontalLayout_3.setObjectName("horizontalLayout_3")
-        self.pushButton = QtWidgets.QPushButton(self.horizontalLayoutWidget_3)
-        self.pushButton.setObjectName("pushButton")
-        self.horizontalLayout_3.addWidget(self.pushButton)
-        self.horizontalLayoutWidget_4 = QtWidgets.QWidget(self.centralwidget)
-        self.horizontalLayoutWidget_4.setGeometry(QtCore.QRect(159, 10, 141, 41))
-        self.horizontalLayoutWidget_4.setObjectName("horizontalLayoutWidget_4")
-        self.horizontalLayout_4 = QtWidgets.QHBoxLayout(self.horizontalLayoutWidget_4)
-        self.horizontalLayout_4.setContentsMargins(0, 0, 0, 0)
-        self.horizontalLayout_4.setObjectName("horizontalLayout_4")
-        self.pushButton_2 = QtWidgets.QPushButton(self.horizontalLayoutWidget_4)
-        self.pushButton_2.setObjectName("pushButton_2")
-        self.horizontalLayout_4.addWidget(self.pushButton_2)
-        self.horizontalLayoutWidget_5 = QtWidgets.QWidget(self.centralwidget)
-        self.horizontalLayoutWidget_5.setGeometry(QtCore.QRect(319, 10, 141, 41))
-        self.horizontalLayoutWidget_5.setObjectName("horizontalLayoutWidget_5")
-        self.horizontalLayout_5 = QtWidgets.QHBoxLayout(self.horizontalLayoutWidget_5)
-        self.horizontalLayout_5.setContentsMargins(0, 0, 0, 0)
-        self.horizontalLayout_5.setObjectName("horizontalLayout_5")
-        self.pushButton_3 = QtWidgets.QPushButton(self.horizontalLayoutWidget_5)
-        self.pushButton_3.setObjectName("pushButton_3")
-        self.horizontalLayout_5.addWidget(self.pushButton_3)
-        self.horizontalLayoutWidget_6 = QtWidgets.QWidget(self.centralwidget)
-        self.horizontalLayoutWidget_6.setGeometry(QtCore.QRect(480, 10, 141, 41))
-        self.horizontalLayoutWidget_6.setObjectName("horizontalLayoutWidget_6")
-        self.horizontalLayout_6 = QtWidgets.QHBoxLayout(self.horizontalLayoutWidget_6)
-        self.horizontalLayout_6.setContentsMargins(0, 0, 0, 0)
-        self.horizontalLayout_6.setObjectName("horizontalLayout_6")
-        self.pushButton_4 = QtWidgets.QPushButton(self.horizontalLayoutWidget_6)
-        self.pushButton_4.setObjectName("pushButton_4")
-        self.horizontalLayout_6.addWidget(self.pushButton_4)
-        self.textBrowser = QtWidgets.QTextBrowser(self.centralwidget)
-        self.textBrowser.setGeometry(QtCore.QRect(10, 550, 471, 41))
-        self.textBrowser.setObjectName("textBrowser")
-        ServerMainWindow.setCentralWidget(self.centralwidget)
-
-        self.retranslateUi(ServerMainWindow)
-        QtCore.QMetaObject.connectSlotsByName(ServerMainWindow)
-
-    def retranslateUi(self, ServerMainWindow):
-        _translate = QtCore.QCoreApplication.translate
-        ServerMainWindow.setWindowTitle(_translate("ServerMainWindow", "Сервер. Главное окно"))
-        self.label.setText(_translate("ServerMainWindow", "ActiveUsers"))
-        self.label_2.setText(_translate("ServerMainWindow", "Info"))
-        self.label_3.setText(_translate("ServerMainWindow", "AD"))
-        self.pushButton.setText(_translate("ServerMainWindow", "Exit"))
-        self.pushButton_2.setText(_translate("ServerMainWindow", "Refresh"))
-        self.pushButton_3.setText(_translate("ServerMainWindow", "ClientHistory"))
-        self.pushButton_4.setText(_translate("ServerMainWindow", "ServerSetup"))
-
-
-if __name__ == '__main__':
-    app = QApplication(sys.argv)
-    w = Ui_ServerMainWindow()
-    w.show()
-    sys.exit(app.exec_())
