+++ mainapp/ui_server_mainwindow_test_run.py
@@ -1,0 +1,20 @@
+"""test start of main window app"""
+
+import sys
+from PyQt5.QtWidgets import QApplication, QMainWindow
+from ui_server_mainwindow_form import UiServerMainWindowForm
+
+
+class ServerWindowMain(QMainWindow):
+    """doc"""
+    def __init__(self):
+        super(ServerWindowMain, self).__init__()
+        self.ui = UiServerMainWindowForm()
+        self.ui.setupUi(self)
+
+
+if __name__ == '__main__':
+    app = QApplication(sys.argv)
+    application = ServerWindowMain()
+    application.show()
+    sys.exit(app.exec_())
